Adds section to check the interaction of SP:Year
To respond to reviewer comments, we were asked to provide an analysis
of the effect of year on each species - therefore, this model was
fit.

However, there is not enough certainty in the estimates to conclude
much - all the CI's include 0. So, I put the block in an if(FALSE)
block to avoid it getting run if not needed.