Adds LOO calculation.
We added the LOO calculation to allow us to compare between models. I
need to verify that the calculation is correct in the morning, but
currently there is very little difference between models, even though
there are large differences between the number of parameters. Need to
dig into the calculation better.